requirements.txt
playwright==1.49.1
anthropic==0.42.0
twilio==9.4.0
apscheduler==3.10.4
sqlalchemy==2.0.36
loguru==0.7.3
click==8.1.7
python-dotenv==1.0.1
pyyaml==6.0.2
rich==13.9.4
flask==3.1.0
flask-ngrok2==0.0.5
